Refactor map and object handling; update player and carrot classes, add gate functionality, and adjust image paths

+++ src/tools/map.h
@@ -25,7 +25,7 @@ public:
     struct Object
     {
         string name;
-        int x, y, width, height;
+        float x, y, width, height;
     };
     struct ObjectGroup
     {
